refactor(Security): Provider, Filter
