+++ thanksbucket/src/main/java/com/thanksbucket/security/authentication/LoginAuthenticationProvider.java
@@ -1,37 +1,13 @@
 package com.thanksbucket.security.authentication;
 
-import com.thanksbucket.security.authentication.userdetails.AuthMemberContext;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.BadCredentialsException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.stereotype.Component;
 
-@Component
-@RequiredArgsConstructor
-public class LoginAuthenticationProvider implements AuthenticationProvider {
-    private final UserDetailsService userDetailsService;
-    private final PasswordEncoder passwordEncoder;
+public class LoginAuthenticationProvider extends DaoAuthenticationProvider {
 
-    @Override
-    public Authentication authenticate(Authentication authentication) throws AuthenticationException {
-        String username = authentication.getName();
-        String password = (String) authentication.getCredentials();
-
-        AuthMemberContext authMemberContext = (AuthMemberContext) userDetailsService.loadUserByUsername(username);
-        if (!passwordEncoder.matches(password, authMemberContext.getPassword())) {
-            throw new BadCredentialsException("비밀번호가 일치하지 않습니다.");
-        }
-
-        return UsernamePasswordAuthenticationToken.authenticated(authMemberContext, authentication.getCredentials(), authMemberContext.getAuthorities());
-    }
-
-    @Override
-    public boolean supports(Class<?> authentication) {
-        return authentication.equals(UsernamePasswordAuthenticationToken.class);
+    public LoginAuthenticationProvider(PasswordEncoder passwordEncoder, UserDetailsService userDetailsService) {
+        super(passwordEncoder);
+        super.setUserDetailsService(userDetailsService);
     }
 }
